+++ src/wrath_header/encrypt.rs
@@ -9,7 +9,7 @@ use std::io::Write;
 /// Encryption part of a [`ServerCrypto`](crate::wrath_header::ServerCrypto).
 ///
 /// Intended to be kept with the writer half of a connection.
-#[derive(Debug)]
+#[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
 pub struct ServerEncrypterHalf {
     encrypt: InnerCrypto,
     server_header: [u8; SERVER_HEADER_MAXIMUM_LENGTH as usize],
@@ -100,7 +100,7 @@ impl ServerEncrypterHalf {
 /// Encryption part of a [`ClientCrypto`](crate::wrath_header::ClientCrypto).
 ///
 /// Intended to be kept with the writer half of a connection.
-#[derive(Debug)]
+#[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
 pub struct ClientEncrypterHalf {
     encrypt: InnerCrypto,
 }